+++ lib/knock/authenticable.rb
@@ -5,7 +5,7 @@ module Knock::Authenticable
     begin
       token = request.headers['Authorization'].split(' ').last
       payload, header = Knock::AuthToken.new(token: token).validate!
-      @current_user = User.find(payload['user_id'])
+      @current_user = Knock.current_user_from_token.call(payload)
     rescue
       head :unauthorized
     end